--- a/compiler/analysis_common.go
+++ b/compiler/analysis_common.go
@@ -11,46 +11,46 @@ import (
 	"github.com/antonmedv/expr/file"
 )
 
-func (c *compiler) countCommonExpr(node ast.Node) {
+func (c *compiler) analyzeCommonExpr(node ast.Node) {
 	switch n := node.(type) {
 	case *ast.NilNode:
-		c.commonCommonNilNode(n)
+		c.analyzeCommonNilNode(n)
 	case *ast.IdentifierNode:
-		c.countCommonIdentifierNode(n)
+		c.analyzeCommonIdentifierNode(n)
 	case *ast.IntegerNode:
-		c.countCommonIntegerNode(n)
+		c.analyzeCommonIntegerNode(n)
 	case *ast.FloatNode:
-		c.countCommonFloatNode(n)
+		c.analyzeCommonFloatNode(n)
 	case *ast.BoolNode:
-		c.countCommonBoolNode(n)
+		c.analyzeCommonBoolNode(n)
 	case *ast.StringNode:
-		c.countCommonStringNode(n)
+		c.analyzeCommonStringNode(n)
 	case *ast.ConstantNode:
-		c.countCommonConstantNode(n)
+		c.analyzeCommonConstantNode(n)
 	case *ast.UnaryNode:
-		c.countCommonUnaryNode(n)
+		c.analyzeCommonUnaryNode(n)
 	case *ast.BinaryNode:
-		c.countCommonBinaryNode(n)
+		c.analyzeCommonBinaryNode(n)
 	case *ast.ChainNode:
-		c.countCommonChainNode(n)
+		c.analyzeCommonChainNode(n)
 	case *ast.MemberNode:
-		c.countCommonMemberNode(n)
+		c.analyzeCommonMemberNode(n)
 	case *ast.SliceNode:
-		c.countCommonSliceNode(n)
+		c.analyzeCommonSliceNode(n)
 	case *ast.CallNode:
 		c.checkCallNode(n)
 	case *ast.BuiltinNode:
-		c.countCommonBuiltinNode(n)
+		c.analyzeCommonBuiltinNode(n)
 	case *ast.ClosureNode:
-		c.countCommonClosureNode(n)
+		c.analyzeCommonClosureNode(n)
 	case *ast.PointerNode:
-		c.countCommonPointerNode(n)
+		c.analyzeCommonPointerNode(n)
 	case *ast.ConditionalNode:
-		c.countCommonConditionalNode(n)
+		c.analyzeCommonConditionalNode(n)
 	case *ast.ArrayNode:
-		c.countCommonArrayNode(n)
+		c.analyzeCommonArrayNode(n)
 	case *ast.MapNode:
-		c.countCommonMapNode(n)
+		c.analyzeCommonMapNode(n)
 	case *ast.PairNode:
 		// do nothing
 	default:
@@ -58,31 +58,31 @@ func (c *compiler) countCommonExpr(node ast.Node) {
 	}
 }
 
-func (c *compiler) commonCommonNilNode(n *ast.NilNode) {
+func (c *compiler) analyzeCommonNilNode(n *ast.NilNode) {
 	n.SetSubExpr("nil")
 }
 
-func (c *compiler) countCommonIdentifierNode(n *ast.IdentifierNode) {
+func (c *compiler) analyzeCommonIdentifierNode(n *ast.IdentifierNode) {
 	n.SetSubExpr(n.Value)
 }
 
-func (c *compiler) countCommonIntegerNode(n *ast.IntegerNode) {
+func (c *compiler) analyzeCommonIntegerNode(n *ast.IntegerNode) {
 	n.SetSubExpr(strconv.FormatInt(int64(n.Value), 10))
 }
 
-func (c *compiler) countCommonFloatNode(n *ast.FloatNode) {
+func (c *compiler) analyzeCommonFloatNode(n *ast.FloatNode) {
 	n.SetSubExpr(strconv.FormatFloat(n.Value, 'f', 10, 64))
 }
 
-func (c *compiler) countCommonBoolNode(n *ast.BoolNode) {
+func (c *compiler) analyzeCommonBoolNode(n *ast.BoolNode) {
 	n.SetSubExpr(strconv.FormatBool(n.Value))
 }
 
-func (c *compiler) countCommonStringNode(n *ast.StringNode) {
+func (c *compiler) analyzeCommonStringNode(n *ast.StringNode) {
 	n.SetSubExpr(strconv.Quote(n.Value))
 }
 
-func (c *compiler) countCommonConstantNode(n *ast.ConstantNode) {
+func (c *compiler) analyzeCommonConstantNode(n *ast.ConstantNode) {
 	var s string
 	switch n.Value.(type) {
 	case string:
@@ -117,8 +117,8 @@ func (c *compiler) countCommonConstantNode(n *ast.ConstantNode) {
 	n.SetSubExpr(s)
 }
 
-func (c *compiler) countCommonUnaryNode(n *ast.UnaryNode) {
-	c.countCommonExpr(n.Node)
+func (c *compiler) analyzeCommonUnaryNode(n *ast.UnaryNode) {
+	c.analyzeCommonExpr(n.Node)
 	buf := strings.Builder{}
 	switch n.Operator {
 	case "+":
@@ -133,9 +133,9 @@ func (c *compiler) countCommonUnaryNode(n *ast.UnaryNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonBinaryNode(n *ast.BinaryNode) {
-	c.countCommonExpr(n.Left)
-	c.countCommonExpr(n.Right)
+func (c *compiler) analyzeCommonBinaryNode(n *ast.BinaryNode) {
+	c.analyzeCommonExpr(n.Left)
+	c.analyzeCommonExpr(n.Right)
 
 	ls := n.Left.SubExpr()
 	rs := n.Right.SubExpr()
@@ -185,17 +185,17 @@ func (c *compiler) countCommonBinaryNode(n *ast.BinaryNode) {
 	case "??", "and", "or", "||", "&&":
 		// do nothing
 	default:
-		c.emitSubExpr(n.SubExpr(), n.Location())
+		c.countCommonExpr(n.SubExpr(), n.Location())
 	}
 }
 
-func (c *compiler) countCommonChainNode(n *ast.ChainNode) {
-	c.countCommonExpr(n.Node)
+func (c *compiler) analyzeCommonChainNode(n *ast.ChainNode) {
+	c.analyzeCommonExpr(n.Node)
 	n.SetSubExpr(n.Node.SubExpr())
 }
 
-func (c *compiler) countCommonMemberNode(n *ast.MemberNode) {
-	c.countCommonExpr(n.Node)
+func (c *compiler) analyzeCommonMemberNode(n *ast.MemberNode) {
+	c.analyzeCommonExpr(n.Node)
 	buf := strings.Builder{}
 	buf.WriteString(n.Node.SubExpr())
 	if n.Method {
@@ -209,7 +209,7 @@ func (c *compiler) countCommonMemberNode(n *ast.MemberNode) {
 			buf.WriteString(".")
 			buf.WriteString(n.Name)
 		} else {
-			c.countCommonExpr(n.Property)
+			c.analyzeCommonExpr(n.Property)
 			buf.WriteString("[")
 			buf.WriteString(n.Property.SubExpr())
 			buf.WriteString("]")
@@ -218,19 +218,19 @@ func (c *compiler) countCommonMemberNode(n *ast.MemberNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonSliceNode(n *ast.SliceNode) {
+func (c *compiler) analyzeCommonSliceNode(n *ast.SliceNode) {
 	buf := strings.Builder{}
-	c.countCommonExpr(n.Node)
+	c.analyzeCommonExpr(n.Node)
 	buf.WriteString(n.Node.SubExpr())
 	buf.WriteString("[")
 	toStr := ""
 	fromStr := ""
 	if n.To != nil {
-		c.countCommonExpr(n.To)
+		c.analyzeCommonExpr(n.To)
 		toStr = n.To.SubExpr()
 	}
 	if n.From != nil {
-		c.countCommonExpr(n.From)
+		c.analyzeCommonExpr(n.From)
 		fromStr = n.From.SubExpr()
 	}
 	buf.WriteString(fromStr)
@@ -242,11 +242,11 @@ func (c *compiler) countCommonSliceNode(n *ast.SliceNode) {
 
 func (c *compiler) checkCallNode(n *ast.CallNode) {
 	buf := strings.Builder{}
-	c.countCommonExpr(n.Callee)
+	c.analyzeCommonExpr(n.Callee)
 	buf.WriteString(n.Callee.SubExpr())
 	buf.WriteString("(")
 	for i, arg := range n.Arguments {
-		c.countCommonExpr(arg)
+		c.analyzeCommonExpr(arg)
 		if i != 0 {
 			buf.WriteString(",")
 		}
@@ -256,12 +256,12 @@ func (c *compiler) checkCallNode(n *ast.CallNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonBuiltinNode(n *ast.BuiltinNode) {
+func (c *compiler) analyzeCommonBuiltinNode(n *ast.BuiltinNode) {
 	buf := strings.Builder{}
 	buf.WriteString(n.Name)
 	buf.WriteString("(")
 	for i, arg := range n.Arguments {
-		c.countCommonExpr(arg)
+		c.analyzeCommonExpr(arg)
 		if i != 0 {
 			buf.WriteString(",")
 		}
@@ -271,21 +271,21 @@ func (c *compiler) countCommonBuiltinNode(n *ast.BuiltinNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonClosureNode(n *ast.ClosureNode) {
-	c.countCommonExpr(n.Node)
+func (c *compiler) analyzeCommonClosureNode(n *ast.ClosureNode) {
+	c.analyzeCommonExpr(n.Node)
 	// omit '{' / '}'
 	n.SetSubExpr(n.Node.SubExpr())
 }
 
-func (c *compiler) countCommonPointerNode(n *ast.PointerNode) {
+func (c *compiler) analyzeCommonPointerNode(n *ast.PointerNode) {
 	// do nothing
 	n.SetSubExpr("#")
 }
 
-func (c *compiler) countCommonConditionalNode(n *ast.ConditionalNode) {
-	c.countCommonExpr(n.Cond)
-	c.countCommonExpr(n.Exp1)
-	c.countCommonExpr(n.Exp2)
+func (c *compiler) analyzeCommonConditionalNode(n *ast.ConditionalNode) {
+	c.analyzeCommonExpr(n.Cond)
+	c.analyzeCommonExpr(n.Exp1)
+	c.analyzeCommonExpr(n.Exp2)
 	buf := strings.Builder{}
 	buf.WriteString(n.Cond.SubExpr())
 	buf.WriteString(" ? ")
@@ -295,11 +295,11 @@ func (c *compiler) countCommonConditionalNode(n *ast.ConditionalNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonArrayNode(n *ast.ArrayNode) {
+func (c *compiler) analyzeCommonArrayNode(n *ast.ArrayNode) {
 	buf := strings.Builder{}
 	buf.WriteString("[")
 	for i, node := range n.Nodes {
-		c.countCommonExpr(node)
+		c.analyzeCommonExpr(node)
 		if i != 0 {
 			buf.WriteString(",")
 		}
@@ -309,12 +309,12 @@ func (c *compiler) countCommonArrayNode(n *ast.ArrayNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) countCommonMapNode(n *ast.MapNode) {
+func (c *compiler) analyzeCommonMapNode(n *ast.MapNode) {
 	pairs := make([]*ast.PairNode, 0)
 	for i := range n.Pairs {
 		pair := n.Pairs[i].(*ast.PairNode)
-		c.countCommonExpr(pair.Key)
-		c.countCommonExpr(pair.Value)
+		c.analyzeCommonExpr(pair.Key)
+		c.analyzeCommonExpr(pair.Value)
 		pairs = append(pairs, pair)
 	}
 	sort.Slice(pairs, func(i, j int) bool {
@@ -334,7 +334,7 @@ func (c *compiler) countCommonMapNode(n *ast.MapNode) {
 	n.SetSubExpr(buf.String())
 }
 
-func (c *compiler) emitSubExpr(subExpr string, loc file.Location) {
+func (c *compiler) countCommonExpr(subExpr string, loc file.Location) {
 	if c.exprRecords == nil || subExpr == "" {
 		return
 	}
@@ -346,8 +346,8 @@ func (c *compiler) emitSubExpr(subExpr string, loc file.Location) {
 	}
 }
 
-func (c *compiler) needCacheCommon(n ast.Node) (bool, int) {
-	needCacheCommon, exprUniqId := false, -1
+func (c *compiler) needReuseCommon(n ast.Node) (bool, int) {
+	needReuseCommon, exprUniqId := false, -1
 	if c.exprRecords != nil {
 		expr := n.SubExpr()
 		hash := fmt.Sprintf("%x", sha1.Sum([]byte(expr)))
@@ -358,9 +358,9 @@ func (c *compiler) needCacheCommon(n ast.Node) (bool, int) {
 				c.commonExpr[cs.id] = expr
 				c.commonExprInc += 1
 			}
-			needCacheCommon = true
+			needReuseCommon = true
 			exprUniqId = cs.id
 		}
 	}
-	return needCacheCommon, exprUniqId
+	return needReuseCommon, exprUniqId
 }