… bench_test.go
… compiler/analysis_common_test.go
… parser/lexer/state.go
… parser/parser_test.go
… vm/opcodes.go
… vm/program.go
… vm/vm.go