I find a tiny improvement for compiler,  which use save result of common sub-expr and used it after.
for  example:
```
(veryHeavyFunc() + tinyFunc1()) + (veryHeavyFunc() + tinyFunc2())
```
we can reuse result of `veryHeavyFunc()` 
I define a new parameter of Config called `AllowReuseCommon`, if it's set true, we will reused result of common expr which has been evaluated before. 
I define three new opcodes for program compiler.
```
OpSaveCommon // save result of common expr to vm.commonCache 
OpLoadCommon // load result of common expr from cache to vm.stack
OpJumpIfSaveCommon // check whether result of common expr has been saved before, If the common expr has already been evaluated and been saved to , it will be evaluated beyond the current common expr. 
```