@antonmedv 2023-07-25
Benchmark results not very promising:
```
goos: linux
goarch: amd64
pkg: github.com/antonmedv/expr
cpu: Intel(R) Xeon(R) Platinum 81[7](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:8)1M CPU @ 2.60GHz
                                       │ /tmp/old.txt │              /tmp/new.txt               │
                                       │    sec/op    │    sec/op     vs base                   │
_expr-2                                   1[8](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:9)6.7n ± 1%    661.0n ± 4%  +254.17% (p=0.000 n=10)
_expr_reuseVm-2                           [9](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:10)2.66n ± 4%   [10](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:11)8.15n ± 4%   +16.72% (p=0.000 n=10)
_filter-2                                 27.30µ ± 4%    25.16µ ± 2%    -7.81% (p=0.000 n=10)
_len-2                                    110.6n ± 3%    567.0n ± 4%  +412.43% (p=0.000 n=10)
_access-2                                 162.2n ± 7%    602.9n ± 3%  +271.56% (p=0.000 n=10)
_accessMap-2                              168.4n ± 2%    625.8n ± 3%  +271.50% (p=0.000 n=10)
_callFunc-2                               956.4n ± 2%   1413.5n ± 5%   +47.79% (p=0.000 n=10)
_callMethod-2                             949.5n ± 2%   1474.5n ± 1%   +55.29% (p=0.000 n=10)
_callField-2                              193.2n ± 3%    690.4n ± 1%  +257.44% (p=0.000 n=10)
_callFast-2                               205.8n ± 3%    705.6n ± 3%  +242.80% (p=0.000 n=10)
_callConstExpr-2                          166.2n ± 2%    646.8n ± 5%  +289.08% (p=0.000 n=10)
_largeStructAccess-2                      369.9n ± 7%    729.3n ± 1%   +97.19% (p=0.000 n=10)
_largeNestedStructAccess-2                373.2n ± 8%    741.6n ± 1%   +98.73% (p=0.000 n=10)
_largeNestedArrayAccess-2                 4.454m ± 3%    4.486m ± 9%         ~ (p=0.353 n=10)
_realWorld-2                              527.0n ± 3%    989.4n ± 4%   +87.77% (p=0.000 n=10)
_realWorld_reuseVm-2                      446.5n ± 3%    435.5n ± 1%    -2.45% (p=0.035 n=10)
_realWorldInsane-2                        215.9µ ± 4%
_realWorldInsane/no_common_reused-2                      198.4µ ± 5%
_realWorldInsane/allow_common_reused-2                   151.2µ ± 6%
geomean                                   925.4n         2.417µ       +115.30%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable

                                       │  /tmp/old.txt  │                /tmp/new.txt                │
                                       │      B/op      │     B/op      vs base                      │
_expr-2                                    32.00 ± 0%     1536.00 ± 0%  +4700.00% (p=0.000 n=10)
_expr_reuseVm-2                            0.000 ± 0%       0.000 ± 0%          ~ (p=1.000 n=10) ¹
_filter-2                                2.617Ki ± 0%     3.141Ki ± 0%    +20.00% (p=0.000 n=10)
_len-2                                     32.00 ± 0%     1536.00 ± 0%  +4700.00% (p=0.000 n=10)
_access-2                                  32.00 ± 0%     1536.00 ± 0%  +4700.00% (p=0.000 n=10)
_accessMap-2                               32.00 ± 0%     1536.00 ± 0%  +4700.00% (p=0.000 n=10)
_callFunc-2                                192.0 ± 0%      1696.0 ± 0%   +783.33% (p=0.000 n=10)
_callMethod-2                              192.0 ± 0%      1696.0 ± 0%   +783.33% (p=0.000 n=10)
_callField-2                               96.00 ± 0%     1600.00 ± 0%  +1566.67% (p=0.000 n=10)
_callFast-2                                96.00 ± 0%     1600.00 ± 0%  +1566.67% (p=0.000 n=10)
_callConstExpr-2                           96.00 ± 0%     1600.00 ± 0%  +1566.67% (p=0.000 n=10)
_largeStructAccess-2                       62.00 ± 2%     1573.00 ± 0%  +2437.10% (p=0.000 n=10)
_largeNestedStructAccess-2                 63.00 ± 2%     1573.00 ± 0%  +2396.83% (p=0.000 n=10)
_largeNestedArrayAccess-2                10.08Mi ± 0%     10.08Mi ± 0%          ~ (p=0.839 n=10)
_realWorld-2                               416.0 ± 0%      1920.0 ± 0%   +361.54% (p=0.000 n=10)
_realWorld_reuseVm-2                       384.0 ± 0%       384.0 ± 0%          ~ (p=1.000 n=10) ¹
_realWorldInsane-2                       16.32Ki ± 1%
_realWorldInsane/no_common_reused-2                       14.95Ki ± 0%
_realWorldInsane/allow_common_reused-2                    28.67Ki ± 0%
geomean                                               ²                  +875.53%                ³ ²
¹ all samples are equal
² summaries must be >0 to compute geomean
³ benchmark set differs from baseline; geomeans may not be comparable

                                       │ /tmp/old.txt │              /tmp/new.txt               │
                                       │  allocs/op   │ allocs/op   vs base                     │
_expr-2                                  1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
_expr_reuseVm-2                          0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
_filter-2                                176.0 ± 0%     172.0 ± 0%    -2.27% (p=0.000 n=10)
_len-2                                   1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
_access-2                                1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
_accessMap-2                             1.000 ± 0%     2.000 ± 0%  +100.00% (p=0.000 n=10)
_callFunc-2                              6.000 ± 0%     7.000 ± 0%   +[16](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:17).67% (p=0.000 n=10)
_callMethod-2                            6.000 ± 0%     7.000 ± 0%   +16.67% (p=0.000 n=10)
_callField-2                             2.000 ± 0%     3.000 ± 0%   +50.00% (p=0.000 n=10)
_callFast-2                              2.000 ± 0%     3.000 ± 0%   +50.00% (p=0.000 n=10)
_callConstExpr-2                         2.000 ± 0%     3.000 ± 0%   +50.00% (p=0.000 n=10)
_largeStructAccess-2                     4.000 ± 0%     5.000 ± 0%   +[25](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:26).00% (p=0.000 n=10)
_largeNestedStructAccess-2               4.000 ± 0%     5.000 ± 0%   +25.00% (p=0.000 n=10)
_largeNestedArrayAccess-2                2.000 ± 0%     3.000 ± 0%   +50.00% (p=0.000 n=10)
_realWorld-2                             2.000 ± 0%     3.000 ± 0%   +50.00% (p=0.000 n=10)
_realWorld_reuseVm-2                     1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
_realWorldInsane-2                       860.0 ± 0%
_realWorldInsane/no_common_reused-2                     840.0 ± 0%
_realWorldInsane/allow_common_reused-2                  567.0 ± 0%
geomean                                             ²                +41.[30](https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393#step:8:31)%                ³ ²
¹ all samples are equal
² summaries must be >0 to compute geomean
³ benchmark set differs from baseline; geomeans may not be comparable
```
https://github.com/antonmedv/expr/actions/runs/5658005304/job/15332440349?pr=393