@antonmedv 2023-07-25
This optimization is actually very interesting to do, but I think it should be done on the compile spec with AST transformation.

But the first step will be to add the variables support in the language.

So for example this expression:

```
foo[0].bar.baz == 2 && foo[0].bar.goz == 3
```

Can be changed into:

```
let tmp = foo[0].bar; tmp.baz == 2 && tmp.goz == 3
```